requirements.txt
python>=3.8
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.0.0
open3d>=0.15.0
plyfile>=0.7.4
pyyaml>=5.4.1
tqdm>=4.62.0
scikit-learn>=1.0.0
matplotlib>=3.4.0